Temporary: Separated view model
