+++ examples/data-objects/prosemirror/src/prosemirror.ts
@@ -3,16 +3,12 @@
  * Licensed under the MIT License.
  */
 
-import { EventEmitter } from "events";
 import {
-    IFluidLoadable,
-    IFluidRouter,
     IRequest,
     IResponse,
     IFluidHandle,
 } from "@fluidframework/core-interfaces";
-import { FluidObjectHandle, FluidDataStoreRuntime } from "@fluidframework/datastore";
-import { ISharedMap, SharedMap } from "@fluidframework/map";
+import { DataObject, DataObjectFactory } from "@fluidframework/aqueduct";
 import {
     IMergeTreeInsertMsg,
     ReferenceType,
@@ -20,13 +16,11 @@ import {
     MergeTreeDeltaType,
     createMap,
 } from "@fluidframework/merge-tree";
-import { IFluidDataStoreContext, IFluidDataStoreFactory } from "@fluidframework/runtime-definitions";
-import { IFluidDataStoreRuntime, IChannelFactory } from "@fluidframework/datastore-definitions";
 import { SharedString } from "@fluidframework/sequence";
-import { IFluidHTMLOptions, IFluidHTMLView } from "@fluidframework/view-interfaces";
-import { EditorView } from "prosemirror-view";
+import { IFluidHTMLView } from "@fluidframework/view-interfaces";
 import { nodeTypeKey } from "./fluidBridge";
 import { FluidCollabManager, IProvideRichTextEditor } from "./fluidCollabManager";
+import {ProseMirrorView} from "./prosemirrorView";
 
 function createTreeMarkerOps(
     treeRangeLabel: string,
@@ -56,79 +50,19 @@ function createTreeMarkerOps(
     ];
 }
 
-class ProseMirrorView implements IFluidHTMLView {
-    private content: HTMLDivElement;
-    private editorView: EditorView;
-    private textArea: HTMLDivElement;
-    public get IFluidHTMLView() { return this; }
-
-    public constructor(private readonly collabManager: FluidCollabManager) { }
-
-    public render(elm: HTMLElement, options?: IFluidHTMLOptions): void {
-        // Create base textarea
-        if (!this.textArea) {
-            this.textArea = document.createElement("div");
-            this.textArea.classList.add("editor");
-            this.content = document.createElement("div");
-            this.content.style.display = "none";
-            this.content.innerHTML = "";
-        }
-
-        // Reparent if needed
-        if (this.textArea.parentElement !== elm) {
-            this.textArea.remove();
-            this.content.remove();
-            elm.appendChild(this.textArea);
-            elm.appendChild(this.content);
-        }
-
-        if (!this.editorView) {
-            this.editorView = this.collabManager.setupEditor(this.textArea);
-        }
-    }
-
-    public remove() {
-        // Maybe implement this some time.
-    }
-}
-
 /**
  * ProseMirror builds a Fluid collaborative text editor on top of the open source text editor ProseMirror.
- * It has its own implementation of IFluidLoadable and does not extend PureDataObject / DataObject. This is
- * done intentionally to serve as an example of exposing the URL and handle via IFluidLoadable.
  */
-export class ProseMirror extends EventEmitter
-    implements IFluidLoadable, IFluidRouter, IFluidHTMLView, IProvideRichTextEditor {
-    public static async load(runtime: IFluidDataStoreRuntime, context: IFluidDataStoreContext) {
-        const collection = new ProseMirror(runtime, context);
-        await collection.initialize();
+export class ProseMirror extends DataObject implements IFluidHTMLView, IProvideRichTextEditor {
 
-        return collection;
-    }
-
-    public get handle(): IFluidHandle<this> { return this.innerHandle; }
-
-    public get IFluidLoadable() { return this; }
-    public get IFluidRouter() { return this; }
     public get IFluidHTMLView() { return this; }
     public get IRichTextEditor() { return this.collabManager; }
 
-    public url: string;
     public text: SharedString;
-    private root: ISharedMap;
     private collabManager: FluidCollabManager;
     private view: ProseMirrorView;
-    private readonly innerHandle: IFluidHandle<this>;
 
-    constructor(
-        private readonly runtime: IFluidDataStoreRuntime,
-        /* Private */ context: IFluidDataStoreContext,
-    ) {
-        super();
-
-        this.url = context.id;
-        this.innerHandle = new FluidObjectHandle(this, this.url, runtime.IFluidHandleContext);
-    }
+    public static get Name() { return "@fluid-example/prosemirror"; }
 
     public async request(request: IRequest): Promise<IResponse> {
         return {
@@ -138,64 +72,47 @@ export class ProseMirror extends EventEmitter
         };
     }
 
-    private async initialize() {
-        if (!this.runtime.existing) {
-            this.root = SharedMap.create(this.runtime, "root");
-            const text = SharedString.create(this.runtime);
+    protected async initializingFirstTime() {
+        const text = SharedString.create(this.runtime);
 
-            const ops = createTreeMarkerOps("prosemirror", 0, 1, "paragraph");
-            text.groupOperation({ ops, type: MergeTreeDeltaType.GROUP });
-            text.insertText(1, "Hello, world!");
+        const ops = createTreeMarkerOps("prosemirror", 0, 1, "paragraph");
+        text.groupOperation({ ops, type: MergeTreeDeltaType.GROUP });
+        text.insertText(1, "Hello, world!");
 
-            this.root.set("text", text.handle);
-            this.root.bindToContext();
-        }
+        this.root.set("text", text.handle);
+    }
 
-        this.root = await this.runtime.getChannel("root") as ISharedMap;
+    protected async hasInitialized() {
         this.text = await this.root.get<IFluidHandle<SharedString>>("text").get();
 
         this.collabManager = new FluidCollabManager(this.text, this.runtime.loader);
+        this.hasValueChanged();
+    }
 
-        // Access for debugging
-        // eslint-disable-next-line dot-notation
-        window["easyComponent"] = this;
+    public hasValueChanged() {
+        this.collabManager?.on("valueChanged", (changed) => {
+            // Here we can set data to original file
+            console.log("something changed ", changed);
+        });
     }
 
     public render(elm: HTMLElement): void {
-        if (!this.view) {
-            this.view = new ProseMirrorView(this.collabManager);
+        if (isWebClient()) {
+            if (!this.view) {
+                this.view = new ProseMirrorView(this.collabManager);
+            }
+            this.view.render(elm);
         }
-        this.view.render(elm);
     }
 }
 
-class ProseMirrorFactory implements IFluidDataStoreFactory {
-    public static readonly type = "@fluid-example/prosemirror";
-    public readonly type = ProseMirrorFactory.type;
+export const ProseMirrorFactory = new DataObjectFactory (
+    ProseMirror.Name,
+    ProseMirror,
+    [SharedString.getFactory()],
+    {},
+);
 
-    public get IFluidDataStoreFactory() { return this; }
-
-    public async instantiateDataStore(context: IFluidDataStoreContext) {
-        const dataTypes = new Map<string, IChannelFactory>();
-        const mapFactory = SharedMap.getFactory();
-        const sequenceFactory = SharedString.getFactory();
-
-        dataTypes.set(mapFactory.type, mapFactory);
-        dataTypes.set(sequenceFactory.type, sequenceFactory);
-
-        const runtime = FluidDataStoreRuntime.load(
-            context,
-            dataTypes,
-        );
-
-        const proseMirrorP = ProseMirror.load(runtime, context);
-        runtime.registerRequestHandler(async (request: IRequest) => {
-            const proseMirror = await proseMirrorP;
-            return proseMirror.request(request);
-        });
-
-        return runtime;
-    }
+const isWebClient = () => {
+    return typeof window !== "undefined" && typeof window.document !== "undefined";
 }
-
-export const fluidExport = new ProseMirrorFactory();
